src/openlrw/src/main/java/unicon/matthews/tenant/Tenant.java
/**
 * 
 */
package unicon.matthews.tenant;

import java.io.Serializable;
import java.util.Map;

import org.apache.commons.lang3.builder.ToStringBuilder;
import org.apache.commons.lang3.builder.ToStringStyle;
import org.springframework.data.annotation.Id;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.databind.annotation.JsonDeserialize;

/**
 * @author ggilbert
 *
 */
@JsonIgnoreProperties(ignoreUnknown=true)
@JsonInclude(JsonInclude.Include.NON_EMPTY)
@JsonDeserialize(builder = Tenant.Builder.class)
public final class Tenant implements Serializable {
  private static final long serialVersionUID = 1L;

  @Id private String id;
  private String name;
  private String description;
  private Map<String, String> metadata;
  
  private Tenant() {}

  public String getId() {
    return id;
  }

  public String getName() {
    return name;
  }

  public String getDescription() {
    return description;
  }
  
  public Map<String, String> getMetadata() {
    return metadata;
  }

  @Override
  public String toString() {
    return ToStringBuilder.reflectionToString(this, ToStringStyle.MULTI_LINE_STYLE);
  }

  @Override
  public int hashCode() {
    final int prime = 31;
    int result = 1;
    result = prime * result + ((description == null) ? 0 : description.hashCode());
    result = prime * result + ((id == null) ? 0 : id.hashCode());
    result = prime * result + ((metadata == null) ? 0 : metadata.hashCode());
    result = prime * result + ((name == null) ? 0 : name.hashCode());
    return result;
  }

  @Override
  public boolean equals(Object obj) {
    if (this == obj)
      return true;
    if (obj == null)
      return false;
    if (getClass() != obj.getClass())
      return false;
    Tenant other = (Tenant) obj;
    if (description == null) {
      if (other.description != null)
        return false;
    } else if (!description.equals(other.description))
      return false;
    if (id == null) {
      if (other.id != null)
        return false;
    } else if (!id.equals(other.id))
      return false;
    if (metadata == null) {
      if (other.metadata != null)
        return false;
    } else if (!metadata.equals(other.metadata))
      return false;
    if (name == null) {
      if (other.name != null)
        return false;
    } else if (!name.equals(other.name))
      return false;
    return true;
  }

  public static class Builder {
    private Tenant _tenant = new Tenant();
    
    public Builder withId(String id) {
      _tenant.id = id;
      return this;
    }
    public Builder withName(String name) {
      _tenant.name = name;
      return this;
    }
    public Builder withDescription(String description) {
      _tenant.description = description;
      return this;
    }
    
    public Builder withMetadata(Map<String, String> metadata) {
      _tenant.metadata = metadata;
      return this;
    }

    public Tenant build() {
      return _tenant;
    }
  }

}
